+++ ProyectoCore/src/Services/ServiceEquipo.cs
@@ -26,17 +26,17 @@ namespace Services
             return _repoEquipo.ObtenerTodos();
         }
 
-        public Equipo? ObtenerPorId(ushort id)
+        public Equipo? ObtenerPorId(byte id)
         {
             return _repoEquipo.ObtenerPorId(id);
         }
 
-        public bool Eliminar(ushort id)
+        public bool Eliminar(byte id)
         {
             return _repoEquipo.Eliminar(id);
         }
 
-        public List<Jugador> ObtenerJugadoresPorEquipo(ushort id)
+        public List<Jugador> ObtenerJugadoresPorEquipo(byte id)
         {
             return _repoEquipo.ObtenerJugadoresPorEquipo(id);
         }
